+++ src/components/layout/layout.test.tsx
@@ -5,7 +5,7 @@ import { create } from 'react-test-renderer';
 
 import { Layout } from './layout';
 
-jest.mock('../../hooks/use-site-metadata', () => {
+jest.mock('../site-metadata/use-site-metadata', () => {
     return {
         useSiteMetadata: jest.fn().mockImplementation(() => {
             return {
