+++ src/components/layout/layout.tsx
@@ -1,26 +1,26 @@
 import React from 'react';
 import type { FunctionComponent, ReactElement, PropsWithChildren } from 'react';
 
-import { useSiteMetadata } from '../../hooks/use-site-metadata';
+import { useSiteMetadata } from '../site-metadata/use-site-metadata';
 
 import { Content } from './content/content';
 import { Footer } from './footer/footer';
 import { Header } from './header/header';
 
 const Layout: FunctionComponent<PropsWithChildren> = ({ children }): ReactElement => {
-    const siteMetadata = useSiteMetadata();
+    const { title, author } = useSiteMetadata();
 
     return (
         <div className="container mx-auto bg-white px-4">
-            <Header title={siteMetadata.title} />
+            <Header title={title} />
 
             <Content>{children}</Content>
 
             <Footer
-                githubLink={siteMetadata.author.social.githubUrl}
-                linkedinLink={siteMetadata.author.social.linkedinUrl}
-                twitterLink={siteMetadata.author.social.twitterUrl}
-                authorName={siteMetadata.author.name}
+                githubLink={author.social.githubUrl}
+                linkedinLink={author.social.linkedinUrl}
+                twitterLink={author.social.twitterUrl}
+                authorName={author.name}
             />
         </div>
     );
